+++ oemer/note_group_extraction.py
@@ -4,7 +4,8 @@ import scipy.ndimage
 
 from oemer import layers
 from oemer.inference import predict
-from oemer.utils import find_closest_staffs, get_global_unit_size, get_unit_size, get_logger
+from oemer.utils import find_closest_staffs, get_global_unit_size, get_unit_size
+from oemer.logging import get_logger
 from oemer.bbox import (
     get_center,
     merge_nearby_bbox,
