+++ oemer/notehead_extraction.py
@@ -7,7 +7,8 @@ import scipy.ndimage
 from oemer import layers
 from oemer.constant import NoteHeadConstant as nhc
 from oemer.bbox import get_bbox, get_center, merge_nearby_bbox, rm_merge_overlap_bbox, to_rgb_img
-from oemer.utils import get_unit_size, find_closest_staffs, get_global_unit_size, get_logger
+from oemer.utils import get_unit_size, find_closest_staffs, get_global_unit_size
+from oemer.logging import get_logger
 from oemer.staffline_extraction import Staff
 from numpy import int64
 from numpy import float64
