+++ oemer/rhythm_extraction.py
@@ -6,7 +6,8 @@ import numpy as np
 import scipy.ndimage
 
 from oemer import layers
-from oemer.utils import get_unit_size, get_logger
+from oemer.utils import get_unit_size
+from oemer.logging import get_logger
 from oemer.bbox import get_center, get_rotated_bbox, to_rgb_img, draw_bounding_boxes
 from oemer.notehead_extraction import NoteType
 from oemer.morph import morph_open, morph_close
